requirements.txt
h5py==3.8.0
imageio==2.31.2
joblib==1.3.1
matplotlib==3.5.3
numpy==1.21.6
opencv_contrib_python==4.8.0.74
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==1.3.5

Pillow==9.5.0
PyQt5==5.15.10
PyQt5_sip==12.13.0
scikit_learn==1.0.2
tensorboardX==2.6.2.2
torch==1.13.0
torchvision==0.14.0
